+++ react/MainImageSwiper.js
@@ -1,0 +1,3 @@
+import MainImageSwiper from './components/ProductImages/components/Carousel/MainImageSwiper'
+
+export default MainImageSwiper